Player Code Refactor
Refactored the player code to have the player manager handle all the calling of functions and generally managing everything about the player.
