+++ Project Soul/Assets/Scripts/PlayerManager.cs
@@ -2,23 +2,54 @@ using PS;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 public class PlayerManager : MonoBehaviour
 {
     InputHandler inputHandler;
     Animator anim;
+    CameraHandler cameraHandler;
+    PlayerLocomotion playerLocomotion;
 
+    public bool isInteracting;
+
+    [Header("Player Flags")]
+    public bool isSprinting;
 
     void Start()
     {
+        cameraHandler = CameraHandler.singleton;
         inputHandler = GetComponent<InputHandler>();
+        playerLocomotion = GetComponent<PlayerLocomotion>();
         anim = GetComponentInChildren<Animator>();
     }
 
     void Update()
     {
-        inputHandler.isInteracting = anim.GetBool("isInteracting");
+        float delta = Time.deltaTime;
+
+        isInteracting = anim.GetBool("isInteracting");
+
+        inputHandler.TickInput(delta);
+        playerLocomotion.HandleMovement(delta);
+        playerLocomotion.HandleRollingAndSprinting(delta);
+    }
+
+    private void FixedUpdate()
+    {
+        float delta = Time.fixedDeltaTime;
+
+        if (cameraHandler != null)
+        {
+            cameraHandler.FollowTarget(delta);
+            cameraHandler.HandleCameraRotation(delta, inputHandler.mouseX, inputHandler.mouseY);
+        }
+    }
+
+    private void LateUpdate()
+    {
         inputHandler.rollFlag = false;
         inputHandler.sprintFlag = false;
+        isSprinting = inputHandler.b_Input;
     }
 }
